Split out functions.
